+++ rustfst/src/algorithms/lazy/cache/first_cache.rs
@@ -64,6 +64,10 @@ impl<W: Semiring, Cache: FstCache<W>> FstCache<W> for FirstCache<W, Cache> {
         self.cache.num_known_states()
     }
 
+    fn compute_num_known_trs(&self) -> usize {
+        self.cache.compute_num_known_trs()
+    }
+
     fn num_trs(&self, id: StateId) -> Option<usize> {
         self.cache.num_trs(id)
     }