+++ rustfst/src/algorithms/lazy/cache/simple_hash_map_cache.rs
@@ -100,6 +100,16 @@ impl<W: Semiring> FstCache<W> for SimpleHashMapCache<W> {
             },
         );
     }
+
+    fn compute_num_known_trs(&self) -> usize {
+        let cached_data = self.trs.lock().unwrap();
+        cached_data
+            .data
+            .values()
+            .map(|it| it.trs.trs().len())
+            .sum()
+    }
+
     fn get_final_weight(&self, id: StateId) -> CacheStatus<Option<W>> {
         match self.final_weights.lock().unwrap().data.get(&id) {
             Some(e) => CacheStatus::Computed(e.clone()),