+++ CHANGELOG.md
@@ -7,8 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.7.4] - 2020-12-10
 
 ## Added
-- `states_range` available on `ExpandedFst` trait
+- `states_range` for `ExpandedFst` trait
 - feature `state-label-u32` is now supported. Label and StateId will be stored as u32 with this feature.
+- `compute_num_known_trs` for `FstCache` trait. 
 
 ## Changed
 - Fix compilation with rust 1.41